feat: fix bug mark all, saved pass, add logout all admin

+++ src/components/Profile/user-security/user-security.component.ts
@@ -2,6 +2,8 @@ import { Subscription } from 'rxjs';
 import { Component, OnInit } from '@angular/core';
 import { Session } from 'models/api.model';
 import { AuthService } from 'services/auth.service';
+import { ConfirmationService } from 'primeng/api';
+import { TranslateService } from '@ngx-translate/core';
 
 @Component({
   selector: 'app-user-security',
@@ -22,6 +24,8 @@ export class UserSecurityComponent implements OnInit {
 
   constructor(
     private authService: AuthService,
+    private confirmationService: ConfirmationService,
+    private translate: TranslateService,
   ) { }
 
   ngOnInit() {
@@ -29,6 +33,7 @@ export class UserSecurityComponent implements OnInit {
   }
 
   getListSession() {
+    this.listSession = [];
     this.isLoading = true;
     this.getListSubscription = this.authService.getAllSessionUser().subscribe(
       (res) => {
@@ -46,13 +51,24 @@ export class UserSecurityComponent implements OnInit {
   }
 
   deleteAll() {
-    this.isLoadingDeleteAll = true;
-    this.deleteAllSubscription = this.authService.deleteAllSessionUser().subscribe(
-      (res) => {
-        this.isLoadingDeleteAll = false;
-        this.getListSession();
+    this.confirmationService.confirm({
+      key: 'logout',
+      header: this.translate.instant('dialog.logoutAll.header'),
+      message: this.translate.instant('dialog.logoutAll.title'),
+      icon: 'pi pi-exclamation-triangle',
+      acceptLabel: this.translate.instant('dialog.logoutAll.yes'),
+      rejectLabel: this.translate.instant('dialog.logoutAll.no'),
+      rejectButtonStyleClass: 'p-button-danger p-button-outlined',
+      accept: () => {
+        this.isLoadingDeleteAll = true;
+        this.deleteAllSubscription = this.authService.deleteAllSessionUser().subscribe(
+          (res) => {
+            this.isLoadingDeleteAll = false;
+            this.getListSession();
+          }
+        );
       }
-    );
+    });
   }
 
   ngOnDestroy() {
